requirements.txt
azure-identity
azure-storage-blob
feedparser
openai>=1.40.0
requests>=2.32.0
PyYAML>=6.0.1
pydub
ffmpeg-python
pandas>=2.2.0
beautifulsoup4
pyarrow>=16.0.0
soccerdata
